+++ src/main/java/com/alejandrom/example/tenpoChallenge/records/HistoryRecordService.java
@@ -16,7 +16,7 @@ public class HistoryRecordService {
         this.historyRecordRepository = historyRecordRepository;
     }
 
-    public void addNewRecord(String endpoint, int num1, int num2, double response){
+    public void addNewRecord(String endpoint, double num1, double num2, double response){
         HistoryRecord historyRecord = new HistoryRecord(LocalDateTime.now(),endpoint,num1,num2,response,"");
         historyRecordRepository.save(historyRecord);
     }
